move math functions from Canvas to Utils
